moved compaction thread interruption to FileCompactor
